ai-api/requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.0
openai>=1.60.0
python-dotenv==1.0.0
pydantic>=2.8.0
python-multipart>=0.0.18